Move embedding-hub endpoint into metabase/embedding/embedding-hub
All consumers are embedding-hub pages and hooks. The file was never
re-exported from the api barrel.
