+++ frontend/src/metabase/embedding/embedding-hub/components/SetupPermissionsAndTenantsPage/SetupPermissionsAndTenantsPage.tsx
@@ -2,8 +2,8 @@ import { useMemo, useRef, useState } from "react";
 import { match } from "ts-pattern";
 import { t } from "ttag";
 
-import { useGetEmbeddingHubChecklistQuery } from "metabase/api/embedding-hub";
 import { Link } from "metabase/common/components/Link";
+import { useGetEmbeddingHubChecklistQuery } from "metabase/embedding/embedding-hub/api/embedding-hub";
 import { OnboardingStepper } from "metabase/embedding/embedding-hub/components/OnboardingStepper";
 import type { OnboardingStepperHandle } from "metabase/embedding/embedding-hub/components/OnboardingStepper/types";
 import {
