+++ frontend/src/metabase/embedding/embedding-hub/hooks/use-completed-embedding-hub-steps.ts
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 
-import { useGetEmbeddingHubChecklistQuery } from "metabase/api/embedding-hub";
+import { useGetEmbeddingHubChecklistQuery } from "metabase/embedding/embedding-hub/api/embedding-hub";
 
 import type { EmbeddingHubStepId } from "../types";
